+++ test/js/components.js
@@ -13,3 +13,5 @@ import '../../weather/current.js';
 import '../../weather/forecast.js';
 import '../../button/share-to.js';
 import '../../krv/events.js';
+import '../../mark-down.js';
+import '../../github/readme.js';